server/pyproject.toml
[tool.poetry]
name = "deals-test-task"
version = "0.1.0"
description = "Test task"
authors = ["Sam <sementip@gmail.com>"]
license = "home"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.5"
psycopg2-binary = "^2.9.8"
djangorestframework = "^3.14.0"
dj-database-url = "^2.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
